+++ app/Modules/ExpertPanel/Events/StepApproved.php
@@ -7,14 +7,14 @@ use Illuminate\Support\Carbon;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Foundation\Events\Dispatchable;
 use App\Modules\ExpertPanel\Models\ExpertPanel;
-use App\Modules\Group\Events\PublishableApplicationEvent;
+use App\Modules\Group\Events\PublishableExpertPanelEvent;
 use Illuminate\Broadcasting\InteractsWithSockets;
-use App\Modules\Group\Events\Traits\IsPublishableApplicationEvent;
+use App\Modules\Group\Events\Traits\IsPublishableExpertPanelEvent;
 
-class StepApproved extends ExpertPanelEvent implements PublishableApplicationEvent
+class StepApproved extends ExpertPanelEvent implements PublishableExpertPanelEvent
 {
     use Dispatchable, InteractsWithSockets, SerializesModels;
-    use IsPublishableApplicationEvent {
+    use IsPublishableExpertPanelEvent {
         getPublishableMessage as protected getBaseMessage;
     }
