+++ app/Modules/ExpertPanel/Events/StepDateApprovedUpdated.php
@@ -5,14 +5,14 @@ namespace App\Modules\ExpertPanel\Events;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Foundation\Events\Dispatchable;
 use App\Modules\ExpertPanel\Models\ExpertPanel;
-use App\Modules\Group\Events\PublishableApplicationEvent;
-use App\Modules\Group\Events\Traits\IsPublishableApplicationEvent;
+use App\Modules\Group\Events\PublishableExpertPanelEvent;
+use App\Modules\Group\Events\Traits\IsPublishableExpertPanelEvent;
 use Illuminate\Broadcasting\InteractsWithSockets;
 
-class StepDateApprovedUpdated extends ExpertPanelEvent implements PublishableApplicationEvent
+class StepDateApprovedUpdated extends ExpertPanelEvent implements PublishableExpertPanelEvent
 {
     use Dispatchable, InteractsWithSockets, SerializesModels;
-    use IsPublishableApplicationEvent {
+    use IsPublishableExpertPanelEvent {
         getPublishableMessage as protected getBaseMessage;
     }
