+++ app/Modules/Group/Events/GroupMemberEvent.php
@@ -2,12 +2,13 @@
 
 namespace App\Modules\Group\Events;
 
-use App\Modules\Group\Events\Traits\IsPublishableApplicationEvent;
+use App\Modules\Group\Events\Traits\IsPublishableGroupEvent;
 use App\Modules\Group\Models\GroupMember;
+use App\Events\PublishableEvent;
 
-abstract class GroupMemberEvent extends GroupEvent implements PublishableApplicationEvent
+abstract class GroupMemberEvent extends GroupEvent implements PublishableEvent
 {
-    use IsPublishableApplicationEvent {
+    use IsPublishableGroupEvent {
         getPublishableMessage as protected getBaseMessage;
     }
 
@@ -38,6 +39,7 @@ abstract class GroupMemberEvent extends GroupEvent implements PublishableApplica
      */
     public function shouldPublish(): bool
     {
+        // FIXME: allow WG events to be published
         return parent::shouldPublish()
             && $this->group->expertPanel->definitionIsApproved;
     }
