+++ app/Modules/Group/Events/PublishableExpertPanelEvent.php
@@ -2,10 +2,9 @@
 
 namespace App\Modules\Group\Events;
 
-use Illuminate\Support\Carbon;
 use App\Events\PublishableEvent;
 
-interface PublishableApplicationEvent extends PublishableEvent
+interface PublishableExpertPanelEvent extends PublishableEvent
 {
     public function mapGeneForMessage($gene): array;
     public function mapMemberForMessage($member): array;
