+++ app/Modules/Group/Events/Traits/IsPublishableGroupEvent.php
@@ -5,7 +5,7 @@ namespace App\Modules\Group\Events\Traits;
 use Illuminate\Support\Str;
 use Illuminate\Support\Carbon;
 
-trait IsPublishableApplicationEvent
+trait IsPublishableGroupEvent
 {
     public function getLogDate(): Carbon
     {
@@ -28,7 +28,7 @@ trait IsPublishableApplicationEvent
                     'type' => $this->group->fullType->name,
             ],
         ];
-        if ($this->group->isEp()) {
+        if ($this->group->isEp()) { // TODO: consider moving this to IsPublisableExpertPanelEvent
             $message['group']['affiliation_id'] = $this->group->expertPanel->affiliation_id;
             $message['group']['scope_description'] = $this->group->expertPanel->scope_description;
             $message['group']['short_name'] = $this->group->expertPanel->short_base_name;
@@ -36,44 +36,10 @@ trait IsPublishableApplicationEvent
             // $message['group']['long_base_name'] = $this->group->expertPanel->long_base_name;
             // $message['group']['hypothesis_group'] = $this->group->expertPanel->hypothesis_group;
             // $message['group']['membership_description'] = $this->group->expertPanel->membership_description;
-
+            if ($this->group->fullType->name === 'vcep') {
+                $message['group']['cspec_url'] = $this->group->expertPanel->affiliation_id;
+            }
         }
-        return [
-            'group' => array_merge(
-                [
-                ],
-                // Conditionally add vcep field below if type is 'vcep'
-                $this->group->fullType->name === 'vcep' ? ['cspec_url' => $this->group->expertPanel->affiliation_id] : [],
-            )
-        ];
-    }
-
-
-    public function mapGeneForMessage($gene): array
-    {
-        $messageGene = [
-            'hgnc_id' => $gene->hgnc_id,
-            'gene_symbol' => $gene->gene_symbol,
-        ];
-
-        if ($gene->mondo_id) {
-            $messageGene['mondo_id'] = $gene->mondo_id;
-            $messageGene['disease_name'] = $gene->disease_name;
-            $messageGene['disease_entity'] = $gene->disease_entity;
-        }
-
-        return $messageGene;
-    }
-
-    public function mapMemberForMessage($member): array
-    {
-        return [
-            'id' => $member->person->uuid,
-            'first_name' => $member->person->first_name,
-            'last_name' => $member->person->last_name,
-            'email' => $member->person->email,
-            'group_roles' => $member->roles->pluck('name')->toArray(),
-            'additional_permissions' => $member->permissions->pluck('name')->toArray(),
-        ];
+        return $message;
     }
 }
