+++ app/Modules/Group/Providers/GroupModuleServiceProvider.php
@@ -21,7 +21,7 @@ use App\Modules\Group\Actions\EventApplicationPublish;
 use App\Modules\Group\Events\ApplicationStepSubmitted;
 use App\Modules\ExpertPanel\Events\ApplicationCompleted;
 use App\Modules\Group\Actions\ApplicationSnapshotCreate;
-use App\Modules\Group\Events\PublishableApplicationEvent;
+use App\Modules\Group\Events\PublishableExpertPanelEvent;
 use App\Modules\Group\Events\ApplicationRevisionsRequested;
 use App\Modules\Group\Actions\ApplicationSubmissionReceiptSend;
 use App\Modules\Group\Actions\NextActionReviewSubmissionComplete;
